rename portfolio lookup router
